refactor: utils/helper 파일 utils 파일로 변경하고 폴더 위치 옮김
